+++ tests/Fixture/Middlewares/EmptyMiddleware.php
@@ -3,7 +3,7 @@
 namespace Rougin\Slytherin\Fixture\Middlewares;
 
 use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 /**
@@ -12,8 +12,19 @@ use Psr\Http\Message\ServerRequestInterface;
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class EmptyMiddleware implements MiddlewareInterface
+class EmptyMiddleware
 {
+    /**
+     * @param  \Psr\Http\Message\ResponseInterface      $request
+     * @param  \Psr\Http\Message\ServerRequestInterface $response
+     * @param  callable|null                            $next
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, $next = null)
+    {
+        return $next($request, $response);
+    }
+
     public function process(ServerRequestInterface $request, DelegateInterface $delegate)
     {
         return $delegate->process($request);