+++ tests/Fixture/Middlewares/SecondMiddleware.php
@@ -2,8 +2,7 @@
 
 namespace Rougin\Slytherin\Fixture\Middlewares;
 
-use Interop\Http\ServerMiddleware\DelegateInterface;
-use Interop\Http\ServerMiddleware\MiddlewareInterface;
+use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 
 /**
@@ -12,14 +11,18 @@ use Psr\Http\Message\ServerRequestInterface;
  * @package Slytherin
  * @author  Rougin Gutib <rougingutib@gmail.com>
  */
-class SecondMiddleware implements MiddlewareInterface
+class SecondMiddleware
 {
-    public function process(ServerRequestInterface $request, DelegateInterface $delegate)
+    /**
+     * @param  \Psr\Http\Message\ResponseInterface      $request
+     * @param  \Psr\Http\Message\ServerRequestInterface $response
+     * @param  callable|null                            $next
+     * @return \Psr\Http\Message\ResponseInterface
+     */
+    public function __invoke(ServerRequestInterface $request, ResponseInterface $response, $next = null)
     {
-        $response = $delegate->process($request);
-
         $response->getBody()->write(' Second!');
 
-        return $response;
+        return $next($request, $response);
     }
 }